Removed old premake functional test
