--- test/functional/toolchains/premake/test_premake.py
@@ -1,124 +1,0 @@
-import os
-import platform
-import textwrap
-
-import pytest
-
-from conan.test.utils.tools import TestClient
-
-
-@pytest.mark.skipif(platform.system() != "Linux", reason="Only for Linux now")
-@pytest.mark.tool("premake")
-def test_premake_args():
-    tc = TestClient(path_with_spaces=False)
-    conanfile = textwrap.dedent(
-        """
-        from conan import ConanFile
-        from conan.tools.premake import Premake, PremakeToolchain
-
-        class Pkg(ConanFile):
-            settings = "os", "compiler", "build_type", "arch"
-            def run(self, cmd, *args, **kwargs):
-                self.output.info(f"Running {cmd}!!")
-            def generate(self):
-                toolchain = PremakeToolchain(self)
-                toolchain.generate()
-            def build(self):
-                premake = Premake(self)
-                premake.luafile = "myproject.lua"
-                premake.arguments = {"myarg": "myvalue"}
-                premake.configure()
-                """
-    )
-    tc.save({"conanfile.py": conanfile})
-    tc.run(
-        "build . -s compiler=msvc -s compiler.version=193 -s compiler.runtime=dynamic"
-    )
-    assert "conanfile.py: Running premake5" in tc.out
-    assert "conanfile.premake5.lua vs2022 --myarg=myvalue!!" in tc.out
-
-
-@pytest.mark.skipif(platform.system() != "Linux", reason="Only for Linux now")
-@pytest.mark.tool("premake")
-def test_premake_full_compilation():
-    tc = TestClient(path_with_spaces=False)
-    tc.run("new cmake_lib -d name=dep -d version=1.0 -o dep")
-
-    consumer_source = textwrap.dedent(
-        """
-        #include <iostream>
-        #include "dep.h"
-
-        int main(void) {
-           dep();
-           std::cout << "Hello World" << std::endl;
-           return 0;
-        }
-    """
-    )
-
-    premake5 = textwrap.dedent(
-        """
-        workspace "Project"
-           language "C++"
-           configurations { "Debug", "Release" }
-
-        project "app"
-           kind "ConsoleApp"
-           files { "**.h", "**.cpp" }
-           filter "configurations:Debug"
-              defines { "DEBUG" }
-              symbols "On"
-           filter "configurations:Release"
-              defines { "NDEBUG" }
-              optimize "On"
-    """
-    )
-
-    conanfile = textwrap.dedent(
-        """
-        from conan import ConanFile
-        from conan.tools.layout import basic_layout
-        from conan.tools.premake import Premake, PremakeDeps, PremakeToolchain
-        import os
-
-        class Pkg(ConanFile):
-            settings = "os", "compiler", "build_type", "arch"
-            name = "pkg"
-            version = "1.0"
-            exports_sources = '*'
-
-            def layout(self):
-                basic_layout(self, src_folder="src")
-
-            def requirements(self):
-                self.requires("dep/1.0")
-
-            def generate(self):
-                deps = PremakeDeps(self)
-                deps.generate()
-                tc = PremakeToolchain(self)
-                tc.generate()
-
-            def build(self):
-                premake = Premake(self)
-                premake.configure()
-                premake.build(workspace="Project", targets=["app"])
-        """
-    )
-
-    tc.save(
-        {
-            "consumer/conanfile.py": conanfile,
-            "consumer/src/hello.cpp": consumer_source,
-            "consumer/src/premake5.lua": premake5,
-        }
-    )
-
-    tc.run("create dep")
-    tc.run("create consumer")
-    bin_folder = os.path.join(tc.created_layout().build(), "build-release", "bin")
-    exe_path = os.path.join(bin_folder, "app")
-    assert os.path.exists(exe_path)
-    tc.run_command(exe_path)
-    assert "dep/1.0: Hello World Release!" in tc.out
