+++ test/functional/toolchains/test_premake.py
@@ -1,74 +1,124 @@
+import os
 import platform
 import textwrap
 
 import pytest
 
-from conan.test.assets.sources import gen_function_cpp
+from conan.test.utils.tools import TestClient
 
 
 @pytest.mark.skipif(platform.system() != "Linux", reason="Only for Linux now")
 @pytest.mark.tool("premake")
-def test_premake(matrix_client):
-    c = matrix_client
-    conanfile = textwrap.dedent("""
-        import os
+def test_premake_args():
+    tc = TestClient(path_with_spaces=False)
+    conanfile = textwrap.dedent(
+        """
         from conan import ConanFile
-        from conan.tools.premake import Premake
-        from conan.tools.microsoft import MSBuild
+        from conan.tools.premake import Premake, PremakeToolchain
+
         class Pkg(ConanFile):
             settings = "os", "compiler", "build_type", "arch"
-            requires = "matrix/1.0"
-            generators = "PremakeDeps", "VCVars"
+            def run(self, cmd, *args, **kwargs):
+                self.output.info(f"Running {cmd}!!")
+            def generate(self):
+                toolchain = PremakeToolchain(self)
+                toolchain.generate()
             def build(self):
-                p = Premake(self)
-                p.configure()
-                build_type = str(self.settings.build_type)
-                if self.settings.os == "Windows":
-                    msbuild = MSBuild(self)
-                    msbuild.build("HelloWorld.sln")
-                else:
-                    self.run(f"make config={build_type.lower()}_x86_64")
-                p = os.path.join(self.build_folder, "bin", build_type, "HelloWorld")
-                self.run(f'"{p}"')
-        """)
-    premake = textwrap.dedent("""
-        -- premake5.lua
+                premake = Premake(self)
+                premake.luafile = "myproject.lua"
+                premake.arguments = {"myarg": "myvalue"}
+                premake.configure()
+                """
+    )
+    tc.save({"conanfile.py": conanfile})
+    tc.run(
+        "build . -s compiler=msvc -s compiler.version=193 -s compiler.runtime=dynamic"
+    )
+    assert "conanfile.py: Running premake5" in tc.out
+    assert "conanfile.premake5.lua vs2022 --myarg=myvalue!!" in tc.out
 
-        include('conandeps.premake5.lua')
 
-        workspace "HelloWorld"
-           conan_setup()
-           configurations { "Debug", "Release" }
-           platforms { "x86_64" }
+@pytest.mark.skipif(platform.system() != "Linux", reason="Only for Linux now")
+@pytest.mark.tool("premake")
+def test_premake_full_compilation():
+    tc = TestClient(path_with_spaces=False)
+    tc.run("new cmake_lib -d name=dep -d version=1.0 -o dep")
 
-        project "HelloWorld"
-           kind "ConsoleApp"
+    consumer_source = textwrap.dedent(
+        """
+        #include <iostream>
+        #include "dep.h"
+
+        int main(void) {
+           dep();
+           std::cout << "Hello World" << std::endl;
+           return 0;
+        }
+    """
+    )
+
+    premake5 = textwrap.dedent(
+        """
+        workspace "Project"
            language "C++"
-           targetdir "bin/%{cfg.buildcfg}"
+           configurations { "Debug", "Release" }
 
+        project "app"
+           kind "ConsoleApp"
            files { "**.h", "**.cpp" }
-
            filter "configurations:Debug"
               defines { "DEBUG" }
               symbols "On"
-
            filter "configurations:Release"
               defines { "NDEBUG" }
               optimize "On"
+    """
+    )
 
-           filter "platforms:x86_64"
-              architecture "x86_64"
-          """)
-    c.save({"conanfile.py": conanfile,
-            "premake5.lua": premake,
-            "main.cpp": gen_function_cpp(name="main", includes=["matrix"], calls=["matrix"])})
-    c.run("build .")
-    assert "main: Release!" in c.out
-    assert "matrix/1.0: Hello World Release!" in c.out
-    if platform.system() == "Windows":
-        assert "main _M_X64 defined" in c.out
-    else:
-        assert "main __x86_64__ defined" in c.out
-    c.run("build . -s build_type=Debug --build=missing")
-    assert "main: Debug!" in c.out
-    assert "matrix/1.0: Hello World Debug!" in c.out
+    conanfile = textwrap.dedent(
+        """
+        from conan import ConanFile
+        from conan.tools.layout import basic_layout
+        from conan.tools.premake import Premake, PremakeDeps, PremakeToolchain
+        import os
+
+        class Pkg(ConanFile):
+            settings = "os", "compiler", "build_type", "arch"
+            name = "pkg"
+            version = "1.0"
+            exports_sources = '*'
+
+            def layout(self):
+                basic_layout(self, src_folder="src")
+
+            def requirements(self):
+                self.requires("dep/1.0")
+
+            def generate(self):
+                deps = PremakeDeps(self)
+                deps.generate()
+                tc = PremakeToolchain(self)
+                tc.generate()
+
+            def build(self):
+                premake = Premake(self)
+                premake.configure()
+                premake.build(workspace="Project", targets=["app"])
+        """
+    )
+
+    tc.save(
+        {
+            "consumer/conanfile.py": conanfile,
+            "consumer/src/hello.cpp": consumer_source,
+            "consumer/src/premake5.lua": premake5,
+        }
+    )
+
+    tc.run("create dep")
+    tc.run("create consumer")
+    bin_folder = os.path.join(tc.created_layout().build(), "build-release", "bin")
+    exe_path = os.path.join(bin_folder, "app")
+    assert os.path.exists(exe_path)
+    tc.run_command(exe_path)
+    assert "dep/1.0: Hello World Release!" in tc.out
